+++ CHANGELOG.md
@@ -6,6 +6,9 @@ v2.0.0-RC.4 (20.02.2021)
 
 - PHP tests use latest signatures (https://github.com/spiral/roadrunner/pull/550)
 - Endure container update to v1.0.0-RC.2 version
+- Remove unneeded mutex from the `http.Workers` method.
+- Rename `checker` plugin package to `status`, remove `/v1` endpoint prefix (#557) 
+- Add static, headers, status, gzip plugins to the `main.go`.
 
 v2.0.0-RC.3 (17.02.2021)
 -------------------